% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAEPirtparams-package.R
\docType{package}
\name{NAEPirtparams-package}
\alias{NAEPirtparams-package}
\title{NAEP IRT Parameter Tables}
\description{
The \code{NAEPirtparams} package provides NAEP IRT parameters, item adjustment information,
and subtest transformation information for NAEP assessments.
}
\details{
The package provides a table called \code{\link{parameters}}, which has item
parameters by year, subject, and grade. The \code{\link{adjustments}} table
provides the item adjustment information that NAEP implemented in scaling (i.e.,
item deletions and collapsed score categories). The table \code{\link{transformations}}
has the location, scale, and weights for the subtests that make up a test.
}
